added vectorized position impl; also planet is flying; need to give grav const and masses

+++ src/logic/Star.java
@@ -3,17 +3,17 @@ package logic;
 import java.awt.Color;
 
 public class Star extends CelestialObject {
-	private int radius;
+	private double radius;
 	private Color color;
 	
-	public Star(int x, int y, int radius) {
+	public Star(double x, double y, double radius) {
 		super(x, y);
 		this.radius = radius;
 		this.color = Color.YELLOW;
 	}
 	
 	public int getRadius() {
-		return radius;
+		return (int)radius;
 	}
 	
 	public Color getColor() {
